H3DAPI warning fixes (vs2010)
Summary: Changes to various things in H3DAPI giving warnings in vs2010

Test Plan: Need to figure one out

Reviewers: markus

Reviewed By: markus

Subscribers: markus, michaels

Tags: #h3dapi_2.x

Differential Revision: https://www.h3dapi.org:4000/D29

+++ trunk/H3DAPI/src/PythonTypes.cpp
@@ -374,7 +374,7 @@ self, deepCopy )" );
                            "Failed to create return value (a list). Please report this issue." );
       return NULL;
     }
-    for( Py_ssize_t i = 0; i < result.result.size(); ++i ) {
+    for( Py_ssize_t i = 0; i < static_cast<Py_ssize_t>( result.result.size() ); ++i ) {
       PyObject *dict = PyDict_New();
       if( !dict ) {
         PyErr_SetString( PyExc_ValueError, 
@@ -467,7 +467,7 @@ self, deepCopy )" );
                            "Failed to create return value (a list). Please report this issue." );
       return NULL;
     }
-    for( Py_ssize_t i = 0; i < result.result.size(); ++i ) {
+    for( Py_ssize_t i = 0; i < static_cast<Py_ssize_t>( result.result.size() ); ++i ) {
       PyObject *dict = PyDict_New();
       if( !dict ) {
         PyErr_SetString( PyExc_ValueError,
